upd member .html compiled.

+++ src/main/java/tw/idv/cha102/g7/member/controller/MemberController.java
@@ -12,8 +12,10 @@ import tw.idv.cha102.g7.member.entity.Member;
 import tw.idv.cha102.g7.member.repo.MemberRepository;
 import tw.idv.cha102.g7.member.service.MemberService;
 
+import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import java.util.List;
 @CrossOrigin
@@ -41,21 +43,6 @@ public class MemberController {
 
 
 
-//    @PostMapping("/login")
-//    public ResponseEntity<String> login(@RequestBody Member member) {
-//        String memEmail = member.getMemEmail();
-//        String memPassword = member.getMemPassword();
-//
-//        String loginResult = memberService.login(memEmail, memPassword);
-//        if ("您已登入團主狀態".equals(loginResult)) {
-//            return ResponseEntity.ok("您已登入團主狀態");
-//        } else if ("您已登入會員狀態".equals(loginResult)) {
-//            return ResponseEntity.ok("您已登入會員狀態");
-//        } else {
-//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("登入失敗");
-//        }
-//    }
-
     /**
      * 會員登入
      * @param loginDTO
@@ -72,6 +59,31 @@ public class MemberController {
         return new ResponseEntity("登入成功",HttpStatus.OK);
     }
 
+    @PostMapping("/logout")
+    public ResponseEntity<String> logout(HttpServletRequest request,
+                                        HttpServletResponse response
+    ) {
+        HttpSession session = request.getSession();
+        String jsessionId = session.getAttribute("memberId").toString();
+        if(jsessionId.isEmpty())
+            return new ResponseEntity("登出失敗",HttpStatus.BAD_REQUEST);
+        session.setAttribute("memberId",null);
+//        if(session.getAttribute("memberId").toString()==null)
+//            System.out.println("NULL");
+        return new ResponseEntity("登出成功",HttpStatus.OK);
+    }
+
+    @RequestMapping("/testlogin")
+    public ResponseEntity<String> testlogin(HttpServletRequest request,
+                                            HttpServletResponse response){
+        HttpSession httpSession = request.getSession();
+        if(httpSession.getAttribute("memberId")==null)
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        String sessionId = httpSession.getAttribute("memberId").toString(); // 保存目前登入的會員id，供後續使用
+        System.out.println(sessionId);
+        return new ResponseEntity<>(sessionId,HttpStatus.OK);
+    }
+
     /**
      * 團主註冊
      * @param member
@@ -83,10 +95,12 @@ public class MemberController {
         String memPassword = member.getMemPassword();
 
         // 檢查用戶是否已登入
-        Integer memSta = memberService.getMemberStatus(memId, memPassword);
-        if (memSta == null || memSta == 0 || memSta == 2) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("請先登入");
-        }
+//        HttpSession httpSession = request.getSession();
+//        String sessionId = httpSession.getAttribute("memberId").toString();
+//        Integer memSta = memberService.getMemberStatus(memId, memPassword);
+//        if (memSta == null || memSta == 0 || memSta == 2) {
+//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("請先登入");
+//        }
 
         // 更新會員的團主狀態為 1（團主）
         member.setMemSta(1);
@@ -103,7 +117,12 @@ public class MemberController {
      * @return
      */
     @DeleteMapping("/delete/{memId}")
-    public ResponseEntity<Boolean> delete(@PathVariable Integer memId) {
+    public ResponseEntity<Boolean> delete(@PathVariable Integer memId,
+                                          HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        String jsessionId = session.getAttribute("memberId").toString();
+        if(jsessionId.isEmpty())
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(false);
         try {
             memberService.deleteById(memId);
             return ResponseEntity.ok(true);
@@ -137,8 +156,14 @@ public class MemberController {
     }
 
 
-
-
+    /**
+     * 信箱驗證
+     * @param member
+     * @return
+     */
+//    public Member sendEmailVerify(Member member) {
+//
+//    }
